+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+- Load the twomemo backend even though SCE is not supported
+
 ## [1.1.0] - 7th of October, 2024
 
 ### Added